CloudRun/Management/code/requirements.txt
google-cloud-logging>=1.15.1
google-api-python-client>=1.12.3
google-cloud-core>=1.4.3
oauth2client>=4.1.3
google-cloud-datastore
google-cloud-scheduler>=2.0.0
flask